+++ internal/interfaces/cli/root.go
@@ -48,6 +48,7 @@ func init() {
 	rootCmd.AddCommand(commands.NewSpecCmd())
 	rootCmd.AddCommand(commands.NewSkillsCmd())
 	rootCmd.AddCommand(commands.NewWorkflowsCmd())
+	rootCmd.AddCommand(commands.NewHooksCmd())
 	rootCmd.AddCommand(commands.NewServeCmd())
 	rootCmd.AddCommand(commands.NewListCmd())