+++ tests/bdd/hook_catalog/hook_catalog_test.go
@@ -1,0 +1,21 @@
+package hook_catalog
+
+import (
+	"testing"
+
+	"github.com/cucumber/godog"
+	"github.com/jorelcb/codify/tests/bdd/commons"
+)
+
+func TestHookCatalogFeature(t *testing.T) {
+	suite := godog.TestSuite{
+		Name:                 "hook_catalog",
+		TestSuiteInitializer: InitializeTestSuite,
+		ScenarioInitializer:  InitializeScenario,
+		Options:              commons.Options("./hook_catalog.feature"),
+	}
+
+	if suite.Run() != 0 {
+		t.Fatal("non-zero status returned, failed to run feature tests")
+	}
+}